Made more classes
