+++ srcs/containers/django/src/pong/static/js/classes/Match.js
@@ -1,0 +1,13 @@
+class Match {
+	constructor(players) {
+		player1 = players[0];
+		player2 = players[1];
+		score = new Score();
+	}
+
+	playMatch() {
+		return result;
+	}
+
+}
+
