+++ srcs/containers/django/src/pong/static/js/classes/Score.js
@@ -1,0 +1,5 @@
+class Score {
+	constructor() {
+		result = [0, 0];
+	}
+}
